requirements.txt
pyTelegramBotAPI
google-generativeai

Flask
google-genai
APScheduler
Pillow
